Add FreeBSD Jenkins CI test script.
